Always update backend hash after merge
We can't detect config drift unless we account for all options passed in
when creating the config. Saving the hash from the config file, then
merging in `-backend-config` options will save a hash in the state
that doesn't match the backend in the state, triggering initialization
each time the same command is run.